requirements.txt
# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
sse-starlette>=2.1.0

# Task queue
celery>=5.4.0
redis>=5.0.4
flower>=2.0.1

# Database
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data processing
pandas>=2.2.2
numpy>=1.26.0
openpyxl>=3.1.2

# LLM
groq>=0.9.0
openai>=1.30.0
tenacity>=8.3.0

# Embeddings
sentence-transformers>=3.0.0
torch>=2.3.0

# Vector store
chromadb>=0.5.0

# PDF generation
weasyprint>=61.0
markdown>=3.6

# Monitoring
prometheus-fastapi-instrumentator>=7.0.0
prometheus-client>=0.20.0
structlog>=24.1.0

# MLflow
mlflow>=2.13.0

# Utils
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
httpx>=0.27.0

# UI (install separately or uncomment)
# streamlit>=1.35.0
# plotly>=5.22.0
